Add projectile shooting for the bow (#7)
* Refactor equipment facing dirty updates

* Rename update-prefixed gameplay helpers

* Add effect entity storage

* Split capacity constants into domain headers

* Hide ammo when equipment has no loaded ammo

* Inline scale entity frame location scaling

* Prepare effect storage for projectile spawning

* Add effect targets and move parent anchor align helper

* Add action state system

* Handle digit 1 release in projectile spawn

* Extract shared pool helpers

* Keep AStar debug rectangles visible

* Track loaded ammo on weapons

* Spawn projectile effects from ammo

* Move size scaling into tilemap load

* Inline load-time scale assignments

* Scale animation frames during tile load

* Add shared group storage helpers

* Load group ids for player and NPC

* Rename group allocation helper

* Store group ids on objects

* Remove unused render group count

* Adjust max groups constant

* Rename projectile spawn system for player shooting

* Assign player group id to spawned effects

* Extract shared entity render buffer fill

* Sort visible render entries after culling

* Render effects above ammo

* Add player depth update system

* Add NPC depth update system

* Add object depth update system

* Sync equipment depth from parent entities

* Use precomputed depth in render buffer sorting

* Sync effect depth from player depth

* Inline parent depth sync assignments

* Track effect parent type for depth sync

* Store effect parent data in SoA form

* Extract render entry compaction and add arrow initialAngle

* Rename Target to AnchorPoint

* Thread rotation center through render buffer

* Aim projectiles from ammo anchor center

* Simplify equipment render buffer fill

* Adjust projectile angle offset

* Add debug rendering for effects

* Reuse entity base debug renderer

* Keep projectile anchor aligned on spawn

* Generalize anchor alignment helper

* Reset copied projectile animation to initial

* Simplify projectile anchor reset flow

* Derive projectile angle from anchor geometry

* Hide debug ammo render when not visible

* Extract anchor center alignment helper

* Mirror projectile geometry before aiming

* Extract entity center world helper

* Clarify projectile anchor variable names

* Add projectile movement system

* Add projectile start anchor point

* Retain shared group on projectile spawn

* Add effect allocation helpers

* Rename effect and group helpers

* Add effect type enum

* Rename projectile movement system

* Rename NPC spawn distance helper

* Add projectile destroy effect system

* Add projectile constants

* Set projectile target past max distance

* Add effect collision id helpers

* Refactor collision id helpers

+++ gg2/structs/collision/CollisionPair.h
@@ -1,8 +1,9 @@
 #pragma once
 #include <cstdint>
+#include "CollisionConstants.h"
 
 struct CollisionPair
 {
-    uint16_t a[MAX_COLLISION_PAIRS];
-    uint16_t b[MAX_COLLISION_PAIRS];
+    uint32_t a[MAX_COLLISION_PAIRS];
+    uint32_t b[MAX_COLLISION_PAIRS];
 };
