+++ gg2/utils/collision/spatialhash/SpatialHashInsertCell.h
@@ -2,7 +2,7 @@
 #include "SpatialHash.h"
 #include "SpatialHashBucketOf.h"
 
-inline void spatialHashInsertCell(SpatialHash &sh, int cx, int cy, uint16_t id)
+inline void spatialHashInsertCell(SpatialHash &sh, int cx, int cy, uint32_t id)
 {
     int b = spatialHashBucketOf(sh, cx, cy);
     if (sh.counts[b] < SpatialHash::MAX_PER_BUCKET)
